Extracted logic to render graph from GraphCard to DefatlChart component
